ai-engine/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
python-multipart==0.0.6
pillow==10.0.0
opencv-python==4.8.0.76
pytesseract==0.3.10
pdf2image==1.16.3
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.34.0
openai==0.28.0
requests>=2.31.0
python-dotenv>=1.0.0
